+++ randomgen/dsfmt.pyx
@@ -290,7 +290,7 @@ cdef class DSFMT(BitGenerator):
         buffered_uniforms = np.empty(DSFMT_N64, dtype=np.double)
         for i in range(DSFMT_N64):
             buffered_uniforms[i] = self.rng_state.buffered_uniforms[i]
-        return {"bit_generator": type(self).__name__,
+        return {"bit_generator": fully_qualified_name(self),
                 "state": {"state": np.asarray(state),
                           "idx": self.rng_state.state.idx},
                 "buffer_loc": self.rng_state.buffer_loc,
@@ -302,7 +302,7 @@ cdef class DSFMT(BitGenerator):
         if not isinstance(value, dict):
             raise TypeError("state must be a dict")
         bitgen = value.get("bit_generator", "")
-        if bitgen != type(self).__name__:
+        if bitgen not in (type(self).__name__, fully_qualified_name(self)):
             raise ValueError("state must be for a {0} "
                              "PRNG".format(type(self).__name__))
         state = check_state_array(value["state"]["state"], 2*DSFMT_N_PLUS_1,