+++ randomgen/mt64.pyx
@@ -190,7 +190,7 @@ cdef class MT64(BitGenerator):
         for i in range(312):
             key[i] = self.rng_state.mt[i]
 
-        return {"bit_generator": type(self).__name__,
+        return {"bit_generator": fully_qualified_name(self),
                 "state": {"key": key, "pos": self.rng_state.mti},
                 "has_uint32": self.rng_state.has_uint32,
                 "uinteger": self.rng_state.uinteger}
@@ -200,7 +200,7 @@ cdef class MT64(BitGenerator):
         if not isinstance(value, dict):
             raise TypeError("state must be a dict")
         bitgen = value.get("bit_generator", "")
-        if bitgen != type(self).__name__:
+        if bitgen not in (type(self).__name__, fully_qualified_name(self)):
             raise ValueError("state must be for a {0} "
                              "PRNG".format(type(self).__name__))
         key = check_state_array(value["state"]["key"], 312, 64, "key")