+++ randomgen/sfc.pyx
@@ -334,7 +334,7 @@ cdef class SFC64(BitGenerator):
             Dictionary containing the information required to describe the
             state of the PRNG
         """
-        return {"bit_generator": type(self).__name__,
+        return {"bit_generator": fully_qualified_name(self),
                 "state": {"a": self.rng_state.a,
                           "b":self.rng_state.b,
                           "c":self.rng_state.c,
@@ -349,7 +349,7 @@ cdef class SFC64(BitGenerator):
         if not isinstance(value, dict):
             raise TypeError("state must be a dict")
         bitgen = value.get("bit_generator", "")
-        if bitgen != type(self).__name__:
+        if bitgen not in (type(self).__name__, fully_qualified_name(self)):
             raise ValueError("state must be for a {0} "
                              "PRNG".format(type(self).__name__))
         self.rng_state.a = value["state"]["a"]