+++ randomgen/speck128.pyx
@@ -290,7 +290,7 @@ cdef class SPECK128(BitGenerator):
             arr[2*i] = self.rng_state.round_key[i].u64[0]
             arr[2*i+1] = self.rng_state.round_key[i].u64[1]
 
-        return {"bit_generator": type(self).__name__,
+        return {"bit_generator": fully_qualified_name(self),
                 "state": {"ctr": ctr,
                           "buffer": buffer.view(np.uint64),
                           "round_key": round_key,
@@ -308,7 +308,7 @@ cdef class SPECK128(BitGenerator):
         if not isinstance(value, dict):
             raise TypeError("state must be a dict")
         bitgen = value.get("bit_generator", "")
-        if bitgen != type(self).__name__:
+        if bitgen not in (type(self).__name__, fully_qualified_name(self)):
             raise ValueError("state must be for a {0} "
                              "PRNG".format(type(self).__name__))