+++ randomgen/xoroshiro128.pyx
@@ -287,7 +287,7 @@ cdef class Xoroshiro128(BitGenerator):
         state = np.empty(2, dtype=np.uint64)
         state[0] = self.rng_state.s[0]
         state[1] = self.rng_state.s[1]
-        return {"bit_generator": type(self).__name__,
+        return {"bit_generator": fully_qualified_name(self),
                 "s": state,
                 "plusplus": self._plusplus,
                 "has_uint32": self.rng_state.has_uint32,
@@ -298,7 +298,7 @@ cdef class Xoroshiro128(BitGenerator):
         if not isinstance(value, dict):
             raise TypeError("state must be a dict")
         bitgen = value.get("bit_generator", "")
-        if bitgen != type(self).__name__:
+        if bitgen not in (type(self).__name__, fully_qualified_name(self)):
             raise ValueError("state must be for a {0} "
                              "PRNG".format(type(self).__name__))
         state = check_state_array(value["s"], 2, 64, "s")