+++ randomgen/xoshiro256.pyx
@@ -268,7 +268,7 @@ cdef class Xoshiro256(BitGenerator):
         state[1] = self.rng_state.s[1]
         state[2] = self.rng_state.s[2]
         state[3] = self.rng_state.s[3]
-        return {"bit_generator": type(self).__name__,
+        return {"bit_generator": fully_qualified_name(self),
                 "s": state,
                 "has_uint32": self.rng_state.has_uint32,
                 "uinteger": self.rng_state.uinteger}
@@ -278,7 +278,7 @@ cdef class Xoshiro256(BitGenerator):
         if not isinstance(value, dict):
             raise TypeError("state must be a dict")
         bitgen = value.get("bit_generator", "")
-        if bitgen != type(self).__name__:
+        if bitgen not in (type(self).__name__, fully_qualified_name(self)):
             raise ValueError("state must be for a {0} "
                              "PRNG".format(type(self).__name__))
         state = check_state_array(value["s"], 4, 64, "s")